refactor(front): centralize types, constants, and env-based API config Extract scattered types into auth.types.ts and profile.types.ts, move API URLs and profile form field configs into constantes.ts, and load the backend proxy target from .env in vite.config.ts. - Remove config/api.ts in favor of constantes.ts - Move CurrentUser, auth payloads, and profile component props to types/ - Externalize VITE_API_PROXY_TARGET for Vite dev proxy (no hardcoded port) - Translate code comments to English and simplify vite-env.d.ts
