+++ app_front-end/src/index.css
@@ -1,45 +1,40 @@
-/**
- * index.css
- * Styles globaux de base pour TROC-SKILLHUB
- */
-
-/* ===== VARIABLES CSS ===== */
+/* ===== CSS VARIABLES ===== */
 :root {
-  /* Couleurs principales */
+  /* Primary colors */
   --color-primary: #008B8B;
   --color-primary-dark: #006666;
   --color-primary-light: #00A5A5;
   
-  /* Couleurs secondaires */
+  /* Secondary colors */
   --color-secondary: #6B4FA0;
   
-  /* Couleurs neutres */
+  /* Neutral colors */
   --color-text: #333333;
   --color-text-light: #666666;
   --color-background: #f5f5f5;
   --color-white: #ffffff;
   --color-border: #e0e0e0;
   
-  /* Couleurs fonctionnelles */
+  /* Functional colors */
   --color-success: #4CAF50;
   --color-warning: #FF9800;
   --color-error: #f44336;
   --color-info: #2196F3;
   
-  /* Typographie */
+  /* Typography */
   font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', 'Roboto', 'Oxygen',
     'Ubuntu', 'Cantarell', 'Fira Sans', 'Droid Sans', 'Helvetica Neue',
     sans-serif;
   line-height: 1.6;
   font-weight: 400;
   
-  /* Optimisation du rendu */
+  /* Rendering optimization */
   font-synthesis: none;
   text-rendering: optimizeLegibility;
   -webkit-font-smoothing: antialiased;
   -moz-osx-font-smoothing: grayscale;
   
-  /* Schéma de couleur (light uniquement) */
+  /* Color scheme (light only) */
   color-scheme: light;
 }
 
@@ -67,7 +62,7 @@ body {
   line-height: inherit;
 }
 
-/* ===== TYPOGRAPHIE ===== */
+/* ===== TYPOGRAPHY ===== */
 h1, h2, h3, h4, h5, h6 {
   color: var(--color-text);
   font-weight: 600;
@@ -104,7 +99,7 @@ p {
   line-height: 1.6;
 }
 
-/* ===== LIENS ===== */
+/* ===== LINKS ===== */
 a {
   color: var(--color-primary);
   text-decoration: none;
@@ -122,7 +117,7 @@ a:focus-visible {
   border-radius: 4px;
 }
 
-/* ===== BOUTONS ===== */
+/* ===== BUTTONS ===== */
 button {
   border-radius: 6px;
   border: 1px solid transparent;
@@ -158,7 +153,7 @@ button:disabled {
   transform: none;
 }
 
-/* Variantes de boutons */
+/* Button variants */
 button.secondary {
   background-color: var(--color-white);
   color: var(--color-primary);
@@ -185,14 +180,14 @@ img {
   display: block;
 }
 
-/* ===== LISTES ===== */
+/* ===== LISTS ===== */
 ul, ol {
   list-style: none;
   padding: 0;
   margin: 0;
 }
 
-/* ===== INPUTS ET FORMULAIRES ===== */
+/* ===== INPUTS AND FORMS ===== */
 input,
 textarea,
 select {
@@ -238,7 +233,7 @@ textarea::placeholder {
   background: #999999;
 }
 
-/* ===== SÉLECTION DE TEXTE ===== */
+/* ===== TEXT SELECTION ===== */
 ::selection {
   background-color: var(--color-primary);
   color: var(--color-white);
@@ -249,8 +244,8 @@ textarea::placeholder {
   color: var(--color-white);
 }
 
-/* ===== ACCESSIBILITÉ ===== */
-/* Désactiver les animations pour ceux qui préfèrent */
+/* ===== ACCESSIBILITY ===== */
+/* Disable animations for users who prefer reduced motion */
 @media (prefers-reduced-motion: reduce) {
   *,
   *::before,
@@ -262,12 +257,12 @@ textarea::placeholder {
   }
 }
 
-/* Focus visible uniquement au clavier (pas à la souris) */
+/* Focus visible on keyboard only (not mouse) */
 *:focus:not(:focus-visible) {
   outline: none;
 }
 
-/* ===== UTILITAIRES ===== */
+/* ===== UTILITIES ===== */
 .sr-only {
   position: absolute;
   width: 1px;
