+++ app_front-end/src/services/knowledgeService.ts
@@ -1,9 +1,8 @@
+import { API_KNOWLEDGES_URL } from "../constantes";
 import type { Knowledge } from "../types/knowledge.types";
 
-const API_BASE_URL = "http://localhost:8099/api";
-
 export const getKnowledges = async (): Promise<Knowledge[]> => {
-  const response = await fetch(`${API_BASE_URL}/knowledges`, {
+  const response = await fetch(API_KNOWLEDGES_URL, {
     credentials: "include",
   });
