+++ app_front-end/src/vite-env.d.ts
@@ -1,12 +1,10 @@
 /// <reference types="vite/client" />
 
-// Déclaration pour les imports CSS
 declare module '*.css' {
   const content: string;
   export default content;
 }
 
-// Déclaration pour les imports d'images
 declare module '*.png' {
   const content: string;
   export default content;
